% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{config_filename}
\alias{config_filename}
\title{Returns the file path to the configuration file.}
\usage{
config_filename()
}
\value{
File path to the configuration fie.
}
\description{
Returns the file path to the configuration file.
}
\examples{
\dontrun{config_filename()}
}
