% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{config_read}
\alias{config_read}
\title{Returns a list with the settings.}
\usage{
config_read()
}
\value{
A list with the settings.
}
\description{
Returns a list with the settings.
}
\examples{
\dontrun{config_read()}
}
