% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{config_write}
\alias{config_write}
\title{Writes a list to the configuration file.}
\usage{
config_write(configList)
}
\arguments{
\item{configList}{A named list.}
}
\value{
Nothing.
}
\description{
Writes a list to the configuration file.
}
\examples{
\dontrun{config_write()}
}
