% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{push_imgur}
\alias{push_imgur}
\title{Pushes current plot to imgur.}
\usage{
push_imgur()
}
\value{
The link to the uploaded image (invisible).
}
\description{
Pushes current plot to imgur.
}
\examples{
\dontrun{push_imgur()}
}
