% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{push_imgur_config}
\alias{push_imgur_config}
\title{Opens a dialog with settings.}
\usage{
push_imgur_config()
}
\value{
Nothing.
}
\description{
Opens a dialog with settings.
}
\examples{
\dontrun{push_imgur_config()}
}
